10.08 delete savetosql add djangosavetosql

+++ BM/initialize_buff_settings.py
@@ -6,8 +6,8 @@ from dotenv import load_dotenv
 dotenv_path = os.path.join(os.path.dirname(__file__), '.env')
 load_dotenv(dotenv_path)
 
-min_price = int(os.environ.get('MIN_PRICE_CNY'))
-max_price = int(os.environ.get('MAX_PRICE_CNY'))
+min_price = float(os.environ.get('MIN_PRICE_CNY'))
+max_price = float(os.environ.get('MAX_PRICE_CNY'))
 if min_price > max_price:
     raise ValueError('Максимальная цена ниже минимальной')
 session = os.environ.get('SESSION')
